Use setUp/tearDown since @Before and @After annotations don't seem to work.

+++ src/de/lmu/ifi/dbs/elki/test/persistent/TestOnDiskArray.java
@@ -5,9 +5,7 @@ import java.io.IOException;
 
 import junit.framework.TestCase;
 
-import org.junit.After;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Test;
 
 import de.lmu.ifi.dbs.elki.persistent.OnDiskArray;
@@ -21,27 +19,29 @@ public class TestOnDiskArray extends TestCase {
   File file = new File("OnDiskArrayTestFile.test.dat");
 
   /**
-   * Fail if test file still exists.
+   * Check that we don't overwrite any file.
    */
-  @Before
-  public void checkTestFile() {
+  @Override
+  protected void setUp() throws Exception {
+    super.setUp();
     if(file.exists()) {
       Assert.fail("Could not run test - test file already exists.");
     }
   }
 
   /**
-   * Clean up afterwards.
+   * Clean up afterwards
    */
-  @After
-  public void removeTestFile() {
-    if(file.exists()) {
+  @Override
+  protected void tearDown() throws Exception {
+    if(file != null && file.exists()) {
       if(!file.delete()) {
         Assert.fail("Error cleaning up: can't remove test file.");
       }
     }
+    super.tearDown();
   }
-
+  
   /**
    * Test the OnDiskArray class.
    * @throws IOException
